refactor: add abstract media library command
