+++ EMS/core-bundle/src/Command/MediaLibrary/MediaLibraryFolderDeleteCommand.php
@@ -4,13 +4,9 @@ declare(strict_types=1);
 
 namespace EMS\CoreBundle\Command\MediaLibrary;
 
-use EMS\CommonBundle\Common\Command\AbstractCommand;
 use EMS\CoreBundle\Command\JobOutput;
 use EMS\CoreBundle\Commands;
-use EMS\CoreBundle\Core\Component\MediaLibrary\Config\MediaLibraryConfig;
-use EMS\CoreBundle\Core\Component\MediaLibrary\Config\MediaLibraryConfigFactory;
 use EMS\CoreBundle\Core\Component\MediaLibrary\Folder\MediaLibraryFolder;
-use EMS\CoreBundle\Core\Component\MediaLibrary\MediaLibraryService;
 use MonorepoBuilderPrefix202311\Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Input\InputArgument;
@@ -22,29 +18,20 @@ use Symfony\Component\Console\Output\OutputInterface;
     description: 'Delete media library folder',
     hidden: false
 )]
-class FolderDeleteCommand extends AbstractCommand
+class MediaLibraryFolderDeleteCommand extends AbstractMediaLibraryCommand
 {
     private MediaLibraryFolder $folder;
     private string $username;
 
     public const ARGUMENT_FOLDER_ID = 'folder-id';
-    public const OPTION_HASH = 'hash';
     public const OPTION_USERNAME = 'username';
-
-    public function __construct(
-        private readonly MediaLibraryConfigFactory $configFactory,
-        private readonly MediaLibraryService $mediaLibraryService
-    ) {
-        parent::__construct();
-    }
 
     protected function configure(): void
     {
+        parent::configure();
         $this
             ->addArgument(self::ARGUMENT_FOLDER_ID, InputArgument::REQUIRED)
-            ->addOption(self::OPTION_HASH, null, InputOption::VALUE_REQUIRED, 'media config hash')
-            ->addOption(self::OPTION_USERNAME, null, InputOption::VALUE_REQUIRED, 'media config hash')
-        ;
+            ->addOption(self::OPTION_USERNAME, null, InputOption::VALUE_REQUIRED, 'media config hash');
     }
 
     protected function initialize(InputInterface $input, OutputInterface $output): void
@@ -52,13 +39,7 @@ class FolderDeleteCommand extends AbstractCommand
         parent::initialize($input, $output);
         $this->io->title('EMS - Media Library - Delete folder');
 
-        $hash = $this->getOptionString(self::OPTION_HASH);
         $folderId = $this->getArgumentString(self::ARGUMENT_FOLDER_ID);
-
-        /** @var MediaLibraryConfig $config */
-        $config = $this->configFactory->createFromHash($hash);
-        $this->mediaLibraryService->setConfig($config);
-
         $this->folder = $this->mediaLibraryService->getFolder($folderId);
         $this->username = $this->getOptionString(self::OPTION_USERNAME);
     }
